filter courseUnits by methods/design agrupar metodos fix login students add new student manually distinguish between myExams and courseExams

+++ app/Http/Controllers/API/LoginController.php
@@ -58,92 +58,101 @@ class LoginController extends Controller
             return response()->json("There was a problem login in. Please try again later",Response::HTTP_INTERNAL_SERVER_ERROR);
         }
         if($response->body()!= "true") {
-            if (Auth::attempt($this->credentials($request))) {
-                $user = Auth::user();
-                $user->refresh();
-            }
-            else
+            if (!Auth::attempt($this->credentials($request))) {
                 return response()->json("Unauthorized.",Response::HTTP_UNAUTHORIZED);
+            }
+            $user = Auth::user();
+            $user->refresh();
         }
         else {
-            //$user = User::where('email', 'like', "%$request->email%")->first();
-
-            if(!$user = User::where('email', 'like', "%$request->email%")->first()){
+//            if(!$user = User::where('email', 'like', "%$request->email%")->first()){
                 //change academicYearCode to 2023/24 format (change accordingly to the webservice argument)
-                $academicYear= AcademicYear::where('selected', true)->first();
-                $academicYearCode = substr($academicYear->code,0,4). "/" .substr($academicYear->code,4,6);
-                if(preg_match('/[0-9]{7}/', $request->email)){
-                    //TODO ANOTHER ISSUE, CHANGING THE SCHOOL YEAR, NEEDS TO REDOO THIS REQUESTS TO GET THE STUDENT COURSES FROM THE YEAR
-                    //TODO problem if student enrroled in 2 or more courses only 1 is saved ... if we do a course enrroled request, we are waisting a lot of time
-                    //            if(!$user = User::where('name', $request->email)->first()){
-                    //TODO webservice/database
-                    $response = Http::get('https://www.dei.estg.ipleiria.pt/servicos/projetos/get_inscricoes_aluno.php?anoletivo='. $academicYearCode .'&num_aluno='.$request->email.'&formato=json');
-                    $student_data = $response->body();
-                    $student_units = json_decode($student_data);
-                    $enrroled_courses = '';
-                    $enrroled_UCS = '';
 
-                    $selectedYear = AcademicYear::where('selected', true)->first();
-                    if($selectedYear){
-                        $activeYear = $selectedYear->id;
-                    } else {
-                        $activeYear = 0;
-                    }
-                    if(!empty($student_units)){
-                        $user= User::create([
-                            "email" => $request->email . '@my.ipleiria.pt',
-                            "name" => $request->email,
-                            "enabled" => true,
-                            "password" => "",
-                        ]);
+            $selectedYear = AcademicYear::where('selected', true)->first();
+            if($selectedYear){
+                $activeYear = $selectedYear->id;
+            } else {
+                $activeYear = 0;
+            }
+            $academicYearCode = substr($selectedYear->code,0,4). "/" .substr($selectedYear->code,4,6);
+            if(preg_match('/[0-9]{7}/', $request->email)){
+                $response = Http::get('https://www.dei.estg.ipleiria.pt/servicos/projetos/get_inscricoes_aluno.php?anoletivo='. $academicYearCode .'&num_aluno='.$request->email.'&formato=json');
+                $student_data = $response->body();
+                $student_units = json_decode($student_data);
+                $enrroled_UCS = '';
 
-                        foreach ($student_units as $unit){
-                            $courseUnit = CourseUnit::where('code', $unit->CD_DISCIP)->where("academic_year_id",$activeYear)->first()->id;
-                            if( !in_array($courseUnit, explode(',',$enrroled_UCS))){
-                                $enrroled_UCS .= ($enrroled_UCS === '' ? '' : ',') . $courseUnit;
-                            }
-                        }
-
-                        $group = Group::isStudent()->first();
-                        $currentGroup = [
-                            'key' => $group->id,
-                            'text' => $group->code,
-                            'value' => $group->name_pt
-                        ];
-
-                        $scopes =  $group->permissions()->where('group_permissions.enabled', true)->groupBy('permissions.code')->pluck('permissions.code')->values()->toArray();
-
-                        $accessToken = $user->createToken('authToken', $scopes)->accessToken;
-
-                        $utils = new Utils();
-                        return response()->json([
-                            'user'          => $user,
-                            'accessToken'   => $accessToken,
-                            'academicYear'  => $utils->getFullYearsAcademicYear($selectedYear ? $selectedYear->display : 0),
-                            'currentGroup'  => $currentGroup
-                        ], Response::HTTP_OK)->withCookie('academic_year', $activeYear )->withCookie('selectedGroup',$group->id)
-                            ->withCookie('courseUnits',$enrroled_UCS);
-//                          ->withCookie('courses', $enrroled_courses)
-                    }
-                    else {
-                        return response()->json("Unauthorized.", Response::HTTP_UNAUTHORIZED);
-                    }
-                }
-                else {
-                    $response = Http::get('https://www.dei.estg.ipleiria.pt/servicos/projetos/get_inscricoes_aluno.php?anoletivo='. $academicYearCode .'&login='.$request->email.'&formato=json');
-                    $docente_data = json_decode($response->body());
-                    $user= User::create([
-                        "email" => $docente_data->email,
-                        "name" => $docente_data->nome,
+                if(!empty($student_units)){
+                    $user= User::firstOrCreate([
+                        "email" => $request->email . '@my.ipleiria.pt',
+                        "name" => $request->email,
                         "enabled" => true,
                         "password" => "",
                     ]);
+
+                    foreach ($student_units as $unit){
+                        $courseUnit = CourseUnit::where('code', $unit->CD_DISCIP)->where("academic_year_id",$activeYear)->first()->id;
+                        if( !in_array($courseUnit, explode(',',$enrroled_UCS))){
+                            $enrroled_UCS .= ($enrroled_UCS === '' ? '' : ',') . $courseUnit;
+                        }
+                    }
+
+                    if($user->groups->count()!=0){
+                        $user->groups()->syncWithoutDetaching(Group::isStudent()->get());
+                    }
+                    $group = Group::isStudent()->first();
+                    $currentGroup = [
+                        'key' => $group->id,
+                        'text' => $group->code,
+                        'value' => $group->name_pt
+                    ];
+
+                    $scopes =  $group->permissions()->where('group_permissions.enabled', true)
+                        ->groupBy('permissions.code')->pluck('permissions.code')->values()->toArray();
+
+                    $accessToken = $user->createToken('authToken', $scopes)->accessToken;
+
+                    $utils = new Utils();
+                    return response()->json([
+                        'user'          => $user,
+                        'accessToken'   => $accessToken,
+                        'academicYear'  => $utils->getFullYearsAcademicYear($selectedYear ? $selectedYear->display : 0),
+                        'currentGroup'  => $currentGroup,
+                        'courseUnits'   =>$enrroled_UCS
+                    ], Response::HTTP_OK)->withCookie('academic_year', $activeYear )->withCookie('selectedGroup',$group->id)
+                        ->withCookie('courseUnits',$enrroled_UCS);
+                }
+                else {
+                    return response()->json("Unauthorized.", Response::HTTP_UNAUTHORIZED);
+                }
+            }
+            else {
+                //TODO change to the correct webservice and fix docente exists validation
+                //Necessario teste pelo prof
+                $user = User::where('email', 'like', "%$request->email%")->first();
+                if(!$user) {
+                    'https://www.dei.estg.ipleiria.pt/servicos/projetos/get_turnos.php?anoletivo='. $selectedYear->code .'&username='. $request->email . '&formato=json';
+                    $response = Http::get('https://www.dei.estg.ipleiria.pt/servicos/projetos/get_inscricoes_aluno.php?anoletivo=' . $academicYearCode . '&login=' . $request->email . '&formato=json');
+                    $docente_data = json_decode($response->body());
+                    $user = User::create([
+                        "email" => $request->email,
+                        "name" => $request->email,
+                        "enabled" => true,
+                        "password" => "",
+                    ]);
+                    $docenteUcs = [];
+                    foreach ($docente_data as $unit){
+                        $courseUnit = CourseUnit::where('code', $unit->codigoUC)->where("academic_year_id",$activeYear)->first()->id;
+                        if( !in_array($courseUnit, $docenteUcs)){
+                            $docenteUcs[] = $courseUnit;
+                            Log::channel('sync_test')->info($docenteUcs);
+                        }
+                    }
+                    $user->courseUnits()->sync($docenteUcs);
                     $user->groups()->syncWithoutDetaching(Group::isTeacher()->get());
                     $user->save();
                 }
             }
         }
-
 
         if (!$user->enabled) {
             return response()->json("Unauthorized.", Response::HTTP_UNAUTHORIZED);
@@ -156,7 +165,6 @@ class LoginController extends Controller
         ];
         $scopes =  $user->groups()->first()->permissions()->where('group_permissions.enabled', true)->groupBy('permissions.code')->pluck('permissions.code')->values()->toArray();
         $accessToken = $user->createToken('authToken', $scopes)->accessToken;
-
         $selectedYear = AcademicYear::where('selected', true)->first();
         if($selectedYear){
             $activeYear = $selectedYear->id;
@@ -177,7 +185,7 @@ class LoginController extends Controller
     public function logout(Request $request)
     {
         $group = Group::where('id', $request->cookie('selectedGroup'))->first()->code;
-        if(str_contains($group,InitialGroups::STUDENT)){
+        if(str_contains($group,InitialGroups::STUDENT) && Auth::guard('api')->user()->groups->count() == 0){
             User::where('email', Auth::guard('api')->user()->email)->forceDelete();
         }
